+++ tortoise/backends/oracle/client.py
@@ -1,18 +1,10 @@
 from __future__ import annotations
 
 import datetime
-import functools
 from typing import TYPE_CHECKING, Any, SupportsInt, cast
 
 import pyodbc
 import pytz
-
-try:
-    from ciso8601 import parse_datetime
-except ImportError:  # pragma: nocoverage
-    from iso8601 import parse_date
-
-    parse_datetime = functools.partial(parse_date, default_timezone=None)
 from pypika_tortoise import OracleQuery
 
 from tortoise.backends.base.client import (
@@ -29,6 +21,7 @@ from tortoise.backends.odbc.client import (
 )
 from tortoise.backends.oracle.executor import OracleExecutor
 from tortoise.backends.oracle.schema_generator import OracleSchemaGenerator
+from tortoise.fields.data import parse_datetime
 
 if TYPE_CHECKING:  # pragma: nocoverage
     import asyncodbc  # pylint: disable=W0611